+++ CHANGELOG.md
@@ -7,6 +7,7 @@ v2.1.1
 
  * ons-select: All the bindings have a wrapper for ons-select and fixed some issues in the core.
  * ons-select: Fix width of the inner element.
+ * ons-switch: Supports 'value' attribute.
 
 v2.1.0
 ----